I moved DatabaseSequence to the entity package and corrected typos
