+++ social-network/backend/java/src/main/java/com/ourproject/socialnetwork/entity/DatabaseSequence.java
@@ -1,4 +1,4 @@
-package com.ourproject.socialnetwork.service;
+package com.ourproject.socialnetwork.entity;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
